Project Updated
This time I took a lot of time to create the dashboard and it experience a lot of moving around. At the end I decided to minimize the clutter and opted for a VBA code to show a graph as needed for the person and get the insight they might need for it.
Later on will be uploading the markdown for the how to make this excel and also the report as if I will be sending this link whit a report to my boss.